src/opensemanticetl/requirements.txt
py2neo==3.1.2
scrapy
warcio
tika
pycurl
rdflib
SPARQLWrapper
requests
pysolr
python-dateutil
lxml
feedparser
celery
pyinotify
twint
dateparser
python-dotenv
